various changes to increase test-coverage
And of course, testing obscure things ends up showing obscure 'bugs' or
better shortcomings/inconsitencies, so lets fix them with the tests.

Git-Dch: Ignore

+++ cmdline/apt-dump-solver.cc
@@ -53,7 +53,8 @@ int main(int argc,const char *argv[])					/*{{{*/
 	   return 0;
 	}
 
-	unlink(filename);
+	if (strcmp(filename, "/dev/null") != 0)
+	   unlink(filename);
 	FileFd input, output;
 	if (input.OpenDescriptor(STDIN_FILENO, FileFd::ReadOnly) == false ||
 	      output.Open(filename, FileFd::WriteOnly | FileFd::Create | FileFd::Exclusive, 0600) == false ||
